Upstream Merge

+++ src/ParkMate/ApplicationServices/Commands/DeleteParkingSpaceCommand.cs
@@ -41,7 +41,12 @@ namespace ParkMate.ApplicationServices.Commands
             DeleteParkingSpaceCommand command,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var parkingSpace = await _repository.GetByIdAsync(command.ParkingSpaceId, command.OwnerId);
+            var parkingSpace = await _repository.GetByIdAsync(command.ParkingSpaceId);
+
+            if(!parkingSpace.OwnerId.Equals(command.OwnerId))
+            {
+                return Result.CommandFail("Not authorized to modify this Parking Space");
+            }
 
             _repository.Delete(parkingSpace);
